DE13817
Error found in US5366: System Settings: Player device settings > Changing default does not update the other tabs set to default in the UI.

The setting is updated but the user has to closer and restart System Settings to see the change.

Set TedE to Use Default.
In the Default tab, changed Player Pin Required from unchecked to checked and clicked Save.
Click on the TedE tab, Player PIN Required is unchecked.

 DE13818
Settings that do not apply to the Ted-E are:

    Allow TV Viewing without purchase
    Wired Network Connection Loss Threshold